+++ src/types/find.ts
@@ -94,7 +94,5 @@ export type ExternalSource =
 
 export interface FindByIdParams {
   externalSource: ExternalSource;
-  includeAdult?: boolean;
   language?: string;
-  page?: number;
 }